hw4-object-detection/pyproject.toml
[project]
name = "hw4-object-detection"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "albumentations>=2.0.5",
    "clip",
    "inference[yolo-world]",
    "opencv-python",
    "opencv-python-headless",
    "pillow",
    "scikit-learn>=1.6.1",
]

[dependency-groups]
dev = [
    "lightning>=2.5.1",
]

[tool.uv.sources]
clip = { git = "https://github.com/openai/CLIP.git" }
